Module Superbol_project.Config

Type definitions

module TYPES : sig ... end

include module type of TYPES with type path = TYPES.path and type config = TYPES.config

type path = TYPES.path =

| RelativeToProjectRoot of string

| RelativeToFileDir of string

type config = TYPES.config = {

mutable cobol_config : Cobol_config.t;

mutable source_format : Cobol_config.source_format_spec;

mutable libpath : path list;

mutable copybook_extensions : string list;

mutable copybook_if_no_extension : bool;

mutable indent_config : (string * int) list;

toml_handle : Ezr_toml.toml_handle;

}

exception ERROR of Superbol_project__.Project_diagnostics.error

type t = config

TOML file management

val new_default : unit -> t

val load_file : ?verbose:bool -> string -> t Cobol_common.Diagnostics.with_diags

load_file ~verbose config_filename loads the given project configuration file. Raises ERROR or Sys_error in case of failure.

val save : ?verbose:bool -> config_filename:string -> t -> unit

val reload : ?verbose:bool -> config_filename:string -> t -> bool Cobol_common.Diagnostics.with_diags

reload ~verbose ~config_filename config returns true, possibly along diagnostics, if the reloaded configuration has changed.

Accessors

val libpath_for : filename:string -> t -> string list

libpath_for ~filename project constructs a list of directory names where copybooks are looked up, for a given source file name, in a project with the given configuration.

val detect_copybook : filename:string -> t -> bool

detect_copybook ~filename config indicates whether a document with the given filename should be treated as a copybook in a project with configuration config.

Intermediate conversion utilities

Those may also raise ERROR

val cobol_config_from_dialect_name : verbose:bool -> string -> Cobol_config.t Cobol_common.Diagnostics.with_diags

val cobol_source_format : string -> Cobol_config.source_format_spec

Cached representation

exception BAD_CHECKSUM

type cached

val to_cache : t -> cached

val of_cache : config_filename:string -> cached -> t