Module Superbol_project.TYPES

type rootdir

type project = {

rootdir : rootdir;

config : Superbol_project__.Project_config.TYPES.config;

config_filename : string;

}

type layout = {

project_config_filename : string;
(*
Name of the TOML file that is to be found at the root of each project's directory tree.
*)

relative_work_dirname : string option;
(*
Relative name of a directory where the LSP should put its working files (caches, etc). No such storage is allowed when None.
*)

rootdir_fallback_policy : rootdir_fallback_policy;
(*
Policy to determine the root directory of projects based on individual filenames.
*)

}

and rootdir_fallback_policy =

| Same_as_file_directory

| Given_directory of string